feat: обработка событий с отзывами

+++ src/main/java/ru/yandex/practicum/filmorate/model/Feed.java
@@ -10,7 +10,7 @@ public class Feed {
     private int userId;
     private int entityId;
 
-    private int timestamp;
+    private long timestamp;
 
     private String eventType;
     private String operation;
